backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.49
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
openai==1.30.1
python-multipart==0.0.9
pydantic==2.12.5
pydantic-settings==2.13.1
email-validator==2.3.0
httpx==0.27.0
bcrypt==4.0.1
fastapi-mail==1.4.1